test: 테스트 생성 및 분할
