+++ src/main/java/domain/deck/DealerDeck.java
@@ -15,7 +15,7 @@ public class DealerDeck extends UserDeck {
     @Override
     public List<Card> getVisibleCards() {
         if (cards.isEmpty()) {
-            throw new IllegalStateException("카드가 존재하지 않는 것은 불가합니다.");
+            throw new IllegalStateException("카드가 존재하지 않을 때 불러오면 안됩니다.");
         }
         return List.of(cards.get(FIRST_INDEX));
     }
